refactor: move the string builder extension methods and associated tests to their own files
