+++ Testing/VelaptorTests/ExtensionMethods/StringBuilderExtensionsTests.cs
@@ -1,0 +1,15 @@
+﻿// <copyright file="StringBuilderExtensionsTests.cs" company="KinsonDigital">
+// Copyright (c) KinsonDigital. All rights reserved.
+// </copyright>
+
+namespace VelaptorTests.ExtensionMethods;
+
+using System.Text;
+
+/// <summary>
+/// Tests the extension methods <see cref="StringBuilder"/> type.
+/// </summary>
+public class StringBuilderExtensionsTests
+{
+    // TODO: Create tests
+}
