+++ Velaptor/UI/TextCursor.cs
@@ -9,6 +9,7 @@ using System.ComponentModel;
 using System.Drawing;
 using System.Numerics;
 using Carbonate.UniDirectional;
+using ExtensionMethods;
 using Graphics;
 using Input;
 using ReactableData;
